+++ packages/extension-koni-ui/src/Popup/Transaction/variants/SendFund.tsx
@@ -5,12 +5,16 @@ import { _AssetRef, _AssetType, _ChainAsset, _ChainInfo, _MultiChainAsset } from
 import { AssetSetting, ExtrinsicType, NotificationType } from '@subwallet/extension-base/background/KoniTypes';
 import { AccountJson } from '@subwallet/extension-base/background/types';
 import { _getXcmUnstableWarning, _isXcmTransferUnstable } from '@subwallet/extension-base/core/substrate/xcm-parser';
-import { _getAssetDecimals, _getOriginChainOfAsset, _getTokenMinAmount, _isAssetFungibleToken, _isChainEvmCompatible, _isMantaZkAsset, _isNativeToken, _isTokenTransferredByEvm } from '@subwallet/extension-base/services/chain-service/utils';
+import { getSnowBridgeGatewayContract } from '@subwallet/extension-base/koni/api/contract-handler/utils';
+import { _getAssetDecimals, _getContractAddressOfToken, _getOriginChainOfAsset, _getTokenMinAmount, _isAssetFungibleToken, _isChainEvmCompatible, _isMantaZkAsset, _isNativeToken, _isTokenTransferredByEvm } from '@subwallet/extension-base/services/chain-service/utils';
 import { SWTransactionResponse } from '@subwallet/extension-base/services/transaction-service/types';
+import { CommonStepType } from '@subwallet/extension-base/types/service-base';
 import { detectTranslate, isSameAddress } from '@subwallet/extension-base/utils';
 import { AccountSelector, AddressInput, AlertBox, AlertModal, AmountInput, ChainSelector, HiddenInput, TokenItemType, TokenSelector } from '@subwallet/extension-koni-ui/components';
-import { useAlert, useFetchChainAssetInfo, useGetChainPrefixBySlug, useHandleSubmitTransaction, useInitValidateTransaction, useIsMantaPayEnabled, useNotification, usePreCheckAction, useRestoreTransaction, useSelector, useSetCurrentPage, useTransactionContext, useWatchTransaction } from '@subwallet/extension-koni-ui/hooks';
-import { getMaxTransfer, makeCrossChainTransfer, makeTransfer } from '@subwallet/extension-koni-ui/messaging';
+import { useAlert, useFetchChainAssetInfo, useGetChainPrefixBySlug, useInitValidateTransaction, useIsMantaPayEnabled, useNotification, usePreCheckAction, useRestoreTransaction, useSelector, useSetCurrentPage, useTransactionContext, useWatchTransaction } from '@subwallet/extension-koni-ui/hooks';
+import useHandleSubmitMultiTransaction from '@subwallet/extension-koni-ui/hooks/transaction/useHandleSubmitMultiTransaction';
+import { approveSpending, getMaxTransfer, getOptimalTransferProcess, makeCrossChainTransfer, makeTransfer } from '@subwallet/extension-koni-ui/messaging';
+import { CommonActionType, commonProcessReducer, DEFAULT_COMMON_PROCESS } from '@subwallet/extension-koni-ui/reducer';
 import { RootState } from '@subwallet/extension-koni-ui/stores';
 import { ChainItemType, FormCallbacks, Theme, ThemeProps, TransferParams } from '@subwallet/extension-koni-ui/types';
 import { findAccountByAddress, findNetworkJsonByGenesisHash, formatBalance, isAccountAll, noop, reformatAddress } from '@subwallet/extension-koni-ui/utils';
@@ -19,7 +23,7 @@ import { Rule } from '@subwallet/react-ui/es/form';
 import BigN from 'bignumber.js';
 import CN from 'classnames';
 import { PaperPlaneRight, PaperPlaneTilt } from 'phosphor-react';
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import React, { useCallback, useEffect, useMemo, useReducer, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 import { useIsFirstRender } from 'usehooks-ts';
@@ -58,6 +62,7 @@ function getTokenItems (
   const isLedger = !!account.isHardware;
   const validGen: string[] = account.availableGenesisHashes || [];
   const validLedgerNetwork = validGen.map((genesisHash) => findNetworkJsonByGenesisHash(chainInfoMap, genesisHash)?.slug);
+  const isGenericLedger = !!account.isGeneric;
   const isAccountEthereum = isEthereumAddress(address);
   const isSetTokenSlug = !!tokenGroupSlug && !!assetRegistry[tokenGroupSlug];
   const isSetMultiChainAssetSlug = !!tokenGroupSlug && !!multiChainAssetMap[tokenGroupSlug];
@@ -68,7 +73,7 @@ function getTokenItems (
     }
 
     const chainAsset = assetRegistry[tokenGroupSlug];
-    const isValidLedger = isLedger ? (isAccountEthereum || validLedgerNetwork.includes(chainAsset?.originChain)) : true;
+    const isValidLedger = isLedger ? (isGenericLedger || validLedgerNetwork.includes(chainAsset?.originChain)) : true;
 
     if (isSetTokenSlug) {
       if (isAssetTypeValid(chainAsset, chainInfoMap, isAccountEthereum) && isValidLedger) {
@@ -91,7 +96,7 @@ function getTokenItems (
   const items: TokenItemType[] = [];
 
   Object.values(assetRegistry).forEach((chainAsset) => {
-    const isValidLedger = isLedger ? (isAccountEthereum || validLedgerNetwork.includes(chainAsset?.originChain)) : true;
+    const isValidLedger = isLedger ? (isGenericLedger || validLedgerNetwork.includes(chainAsset?.originChain)) : true;
     const isTokenFungible = _isAssetFungibleToken(chainAsset);
 
     if (!(isTokenFungible && isAssetTypeValid(chainAsset, chainInfoMap, isAccountEthereum) && isValidLedger)) {
@@ -253,12 +258,14 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
   const [forceUpdateMaxValue, setForceUpdateMaxValue] = useState<object|undefined>(undefined);
   const chainStatus = useMemo(() => chainStatusMap[chain]?.connectionStatus, [chain, chainStatusMap]);
 
+  const [processState, dispatchProcessState] = useReducer(commonProcessReducer, DEFAULT_COMMON_PROCESS);
+
   const handleTransferAll = useCallback((value: boolean) => {
     setForceUpdateMaxValue({});
     setIsTransferAll(value);
   }, []);
 
-  const { onError, onSuccess } = useHandleSubmitTransaction(handleTransferAll);
+  const { onError, onSuccess } = useHandleSubmitMultiTransaction(dispatchProcessState, handleTransferAll);
 
   const destChainItems = useMemo<ChainItemType[]>(() => {
     return getTokenAvailableDestinations(asset, xcmRefMap, chainInfoMap);
@@ -366,7 +373,7 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
       const destChainInfo = chainInfoMap[destChain];
       const availableGen: string[] = account.availableGenesisHashes || [];
 
-      if (!isEthereumAddress(account.address) && !availableGen.includes(destChainInfo?.substrateInfo?.genesisHash || '')) {
+      if (!account.isGeneric && !availableGen.includes(destChainInfo?.substrateInfo?.genesisHash || '')) {
         const destChainName = destChainInfo?.name || 'Unknown';
 
         return Promise.reject(t('Wrong network. Your Ledger account is not supported by {{network}}. Please choose another receiving account and try again.', { replace: { network: destChainName } }));
@@ -446,12 +453,9 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
     [form, assetRegistry, isTransferAll, persistData]
   );
 
-  // Submit transaction
-  const doSubmit: FormCallbacks<TransferParams>['onFinish'] = useCallback((values: TransferParams) => {
+  const isShowWarningOnSubmit = useCallback((values: TransferParams): boolean => {
     setLoading(true);
-    const { asset, chain, destChain, from: _from, to, value } = values;
-
-    let sendPromise: Promise<SWTransactionResponse>;
+    const { asset, chain, destChain, from: _from } = values;
 
     const account = findAccountByAddress(accounts, _from);
 
@@ -462,12 +466,8 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
         type: 'error'
       });
 
-      return;
+      return true;
     }
-
-    const chainInfo = chainInfoMap[chain];
-    const addressPrefix = chainInfo?.substrateInfo?.addressPrefix ?? 42;
-    const from = reformatAddress(_from, addressPrefix);
 
     const isLedger = !!account.isHardware;
     const isEthereum = isEthereumAddress(account.address);
@@ -483,11 +483,35 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
               type: 'warning'
             });
 
-            return;
+            return true;
           }
         }
       }
+    } else {
+      if (isLedger) {
+        setLoading(false);
+        notification({
+          message: t('This feature is not available for Ledger account'),
+          type: 'warning'
+        });
 
+        return true;
+      }
+    }
+
+    return false;
+  }, [accounts, assetRegistry, notification, t]);
+
+  const handleBasicSubmit = useCallback((values: TransferParams): Promise<SWTransactionResponse> => {
+    const { asset, chain, destChain, from: _from, to, value } = values;
+
+    let sendPromise: Promise<SWTransactionResponse>;
+
+    const chainInfo = chainInfoMap[chain];
+    const addressPrefix = chainInfo?.substrateInfo?.addressPrefix ?? 42;
+    const from = reformatAddress(_from, addressPrefix);
+
+    if (chain === destChain) {
       // Transfer token or send fund
       sendPromise = makeTransfer({
         from,
@@ -498,16 +522,6 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
         transferAll: isTransferAll
       });
     } else {
-      if (isLedger) {
-        setLoading(false);
-        notification({
-          message: t('This feature is not available for Ledger account'),
-          type: 'warning'
-        });
-
-        return;
-      }
-
       // Make cross chain transfer
       sendPromise = makeCrossChainTransfer({
         destinationNetworkKey: destChain,
@@ -520,17 +534,80 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
       });
     }
 
+    return sendPromise;
+  }, [chainInfoMap, isTransferAll]);
+
+  // todo: must refactor later, temporary solution to support SnowBridge
+  const handleSnowBridgeSpendingApproval = useCallback((values: TransferParams): Promise<SWTransactionResponse> => {
+    const tokenInfo = assetRegistry[values.asset];
+
+    return approveSpending({
+      amount: values.value,
+      contractAddress: _getContractAddressOfToken(tokenInfo),
+      spenderAddress: getSnowBridgeGatewayContract(values.chain),
+      chain: values.chain,
+      owner: values.from
+    });
+  }, [assetRegistry]);
+
+  // Submit transaction
+  const doSubmit: FormCallbacks<TransferParams>['onFinish'] = useCallback((values: TransferParams) => {
+    if (isShowWarningOnSubmit(values)) {
+      return;
+    }
+
+    const submitData = async (step: number): Promise<boolean> => {
+      dispatchProcessState({
+        type: CommonActionType.STEP_SUBMIT,
+        payload: null
+      });
+
+      const isFirstStep = step === 0;
+      const isLastStep = step === processState.steps.length - 1;
+      const needRollback = step === 1;
+
+      try {
+        if (isFirstStep) {
+          // todo: validate process
+          dispatchProcessState({
+            type: CommonActionType.STEP_COMPLETE,
+            payload: true
+          });
+          dispatchProcessState({
+            type: CommonActionType.STEP_SUBMIT,
+            payload: null
+          });
+
+          return await submitData(step + 1);
+        } else {
+          const stepType = processState.steps[step].type;
+          const submitPromise: Promise<SWTransactionResponse> | undefined = stepType === CommonStepType.TOKEN_APPROVAL ? handleSnowBridgeSpendingApproval(values) : handleBasicSubmit(values);
+
+          const rs = await submitPromise;
+          const success = onSuccess(isLastStep, needRollback)(rs);
+
+          if (success) {
+            return await submitData(step + 1);
+          } else {
+            return false;
+          }
+        }
+      } catch (e) {
+        onError(e as Error);
+
+        return false;
+      }
+    };
+
     setTimeout(() => {
       // Handle transfer action
-      sendPromise
-        .then(onSuccess)
+      submitData(processState.currentStep)
         .catch(onError)
         .finally(() => {
           setLoading(false);
-        })
-      ;
+        });
     }, 300);
-  }, [accounts, chainInfoMap, assetRegistry, notification, t, isTransferAll, onSuccess, onError]);
+  }, [handleBasicSubmit, handleSnowBridgeSpendingApproval, isShowWarningOnSubmit, onError, onSuccess, processState.currentStep, processState.steps]);
 
   const onFilterAccountFunc = useMemo(() => filterAccountFunc(chainInfoMap, assetRegistry, multiChainAssetMap, sendFundSlug), [assetRegistry, chainInfoMap, multiChainAssetMap, sendFundSlug]);
 
@@ -550,7 +627,7 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
       if (_isXcmTransferUnstable(originChainInfo, destChainInfo)) {
         openAlert({
           type: NotificationType.WARNING,
-          content: t(_getXcmUnstableWarning(originChainInfo)),
+          content: t(_getXcmUnstableWarning(originChainInfo, destChainInfo)),
           title: t('Pay attention!'),
           okButton: {
             text: t('Continue'),
@@ -694,6 +771,28 @@ const _SendFund = ({ className = '' }: Props): React.ReactElement<Props> => {
       setIsTransferAll(true);
     }
   }, [maxTransfer, transferAmount]);
+
+  useEffect(() => {
+    getOptimalTransferProcess({
+      amount: transferAmount,
+      address: from,
+      originChain: chain,
+      tokenSlug: asset,
+      destChain
+    })
+      .then((result) => {
+        dispatchProcessState({
+          payload: {
+            steps: result.steps,
+            feeStructure: result.totalFee
+          },
+          type: CommonActionType.STEP_CREATE
+        });
+      })
+      .catch((e) => {
+        console.log('error', e);
+      });
+  }, [asset, chain, destChain, from, transferAmount]);
 
   useRestoreTransaction(form);
   useInitValidateTransaction(validateFields, form, defaultData);